Update suffix finding and replacement methods so that they operate on and return slices of runes.
There were many places in which I was finding a string suffix and then looking at its `len`, which was clearly incorrect because I needed instead a rune count.  Thus, the suffix finding methods really need to look for string suffixes and return the found suffix and a slice of runes for the found suffix.  That slice of runes can be used for further operations using the found suffix.  Clearly, this didn't matter for English because the Unicode code points are the same as ASCII.